pyproject.toml
[tool.poetry]
name = "echo-karriere-backend"
version = "0.1.0"
description = "A Django GraphQL backend for www.echokarriere.no"
authors = ["Sondre Nilsen <nilsen.sondre@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.0.6"
django-filter = "^2.2.0"
django-environ = "^0.4.5"
django-ckeditor = "^5.9.0"
Pillow = "^7.1.2"
django-cors-headers = "^3.3.0"
psycopg2 = { version = "^2.8", optional = true }
graphene-django = "^2.10.1"
django-graphql-jwt = "0.3.0"
django-graphql-auth = "^0.3.10"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pre-commit = "^2.4.0"
flake8 = "^3.8.1"
flake8-django = "^1.0.0"
ipython = "^7.14.0"
coverage = {extras = ["toml"], version = "^5.1"}
isort = {version = "^4.3.21", extras = ["pyproject"]}
seed-isort-config = "^2.1.1"
flake8-isort = "^3.0.0"
psycopg2-binary = "^2.8.5"

[tool.poetry.extras]
production = ["psycopg2"]

[tool.black]
line_length = 120
target-version = ['py37', 'py38']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
line_length = 120
skip_glob = "*/migrations/*"
include_trailing_comma = true
known_django = "django"
sections=["FUTURE","STDLIB","DJANGO","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
known_third_party = ["ckeditor", "django", "environ", "graphene", "graphene_django", "graphql_auth"]

[tool.coverage.run]
source = [ "." ]
